+++ src/Config/Rector/Downgrade/Configurators/AbstractPluginDowngradeContainerConfigurationService.php
@@ -6,7 +6,7 @@ namespace PoP\PoP\Config\Rector\Downgrade\Configurators;
 
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
-abstract class AbstractPluginContainerConfigurationService extends AbstractContainerConfigurationService
+abstract class AbstractPluginDowngradeContainerConfigurationService extends AbstractDowngradeContainerConfigurationService
 {
     protected string $pluginDir;
