+++ src/Config/Rector/Downgrade/Configurators/MonorepoContainerConfigurationService.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PoP\PoP\Config\Rector\Downgrade\Configurators;
 
-class MonorepoContainerConfigurationService extends AbstractContainerConfigurationService
+class MonorepoContainerConfigurationService extends AbstractDowngradeContainerConfigurationService
 {
     /**
      * @return string[]
